Restrict activity creation to admins
