+++ src/app/activities/new/form.tsx
@@ -1,0 +1,60 @@
+'use client';
+
+import { useState } from 'react';
+import { useRouter } from 'next/navigation';
+import { Button } from '@/components/ui/button';
+
+export default function CreateActivityForm() {
+  const [name, setName] = useState('');
+  const [date, setDate] = useState('');
+  const [image, setImage] = useState('');
+  const [description, setDescription] = useState('');
+  const router = useRouter();
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    await fetch('/api/activities', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ name, date, image, description }),
+    });
+    setName('');
+    setDate('');
+    setImage('');
+    setDescription('');
+    router.push('/activities');
+    router.refresh();
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="space-y-4">
+      <input
+        type="text"
+        placeholder="Nombre de la actividad"
+        value={name}
+        onChange={(e) => setName(e.target.value)}
+        className="w-full border px-2 py-1"
+      />
+      <input
+        type="date"
+        value={date}
+        onChange={(e) => setDate(e.target.value)}
+        className="w-full border px-2 py-1"
+      />
+      <input
+        type="url"
+        placeholder="URL de la imagen"
+        value={image}
+        onChange={(e) => setImage(e.target.value)}
+        className="w-full border px-2 py-1"
+      />
+      <textarea
+        placeholder="Descripción"
+        value={description}
+        onChange={(e) => setDescription(e.target.value)}
+        className="w-full border px-2 py-1"
+      />
+      <Button type="submit">Guardar</Button>
+    </form>
+  );
+}
